Clamp the capacity hint before narrowing it, add a spawnSync spread test
Bun.spawnSync builds its result with createEmptyObject(global, 0), so
spreading the result hit the same hasInlineStorage() assertion. Add a
regression test for that surface.

Clamp the size_t hint to maxInlineCapacity before the cast to unsigned
so a hint that is a multiple of 2^32 cannot narrow to zero and recreate
the zero-capacity structure.

+++ src/jsc/bindings/bindings.cpp
@@ -2694,7 +2694,7 @@ JSC__JSObject__create(JSC::JSGlobalObject* globalObject, size_t initialCapacity,
     void (*ArgFn3)(void* arg0, JSC::JSObject* arg1, JSC::JSGlobalObject* arg2))
 {
     JSC::JSObject* object = initialCapacity
-        ? JSC::constructEmptyObject(globalObject, globalObject->objectPrototype(), std::min(static_cast<unsigned>(initialCapacity), JSFinalObject::maxInlineCapacity))
+        ? JSC::constructEmptyObject(globalObject, globalObject->objectPrototype(), static_cast<unsigned>(std::min(initialCapacity, static_cast<size_t>(JSFinalObject::maxInlineCapacity))))
         : JSC::constructEmptyObject(globalObject);
 
     ArgFn3(arg2, object, globalObject);
@@ -2729,7 +2729,7 @@ JSC::EncodedJSValue JSC__JSValue__createEmptyObject(JSC::JSGlobalObject* globalO
     if (!initialCapacity)
         return JSC::JSValue::encode(JSC::constructEmptyObject(globalObject));
     return JSC::JSValue::encode(
-        JSC::constructEmptyObject(globalObject, globalObject->objectPrototype(), std::min(static_cast<unsigned int>(initialCapacity), JSFinalObject::maxInlineCapacity)));
+        JSC::constructEmptyObject(globalObject, globalObject->objectPrototype(), static_cast<unsigned>(std::min(initialCapacity, static_cast<size_t>(JSFinalObject::maxInlineCapacity)))));
 }
 
 extern "C" uint64_t Bun__Blob__getSizeForBindings(void* blob);
